Remove null checks around EventDispatcher
Originally the EventDispatcher was added with the intent that it could
be nullable, however a disconnect in how it was coded in the
ExceptionListener and how it was being injected left the code in a sort
of paradox with itself: we were always requiring it, but we were
null-checking it in the code as if it were possible to leave the
dispatcher null. Ultimately, after a few other changes and suggestions
from other participants in the code review it made the most sense just
to require the EventDispatcher and, within the code, make the
assumption that it would always be there rather than checking to see if
there was a null value.