fix(types): added nullability to getDatabase method
If getDatabase doesn't allow null return values, it will fail when the Driver returns null for this method.

It happens in https://github.com/webmozart/doctrine-dbal/blob/master/lib/Doctrine/DBAL/Driver/PDOSqlite/Driver.php#L114 when using `DATABASE_URL="sqlite:///:memory:"`

this is weird because in-memory dbs are usually used in a test environments were sentry is deactivated, but this class is still being loaded and throws errors in all the tests. 

I get the following error in all the tests:
`TypeError : Return value of Sentry\SentryBundle\Tracing\Doctrine\DBAL\TracingDriver::getDatabase() must be of the type string, null returned`

with this change it works for me.