fix: `symfony/yaml` is a required dependency
getsentry/sentry-symfony#968 migrated the service definitions from XML to YAML, however, the bundle only declares the YAML component as a development dependency.  This causes the bundle's services to not be loaded if an application does not have the component installed, therefore, the YAML component is now a production dependency of the bundle.